Refactor commands into classes
