+++ src/main/java/command/ByeCommand.java
@@ -1,0 +1,11 @@
+package command;
+
+import exception.DukeException;
+
+public class ByeCommand extends Command {
+    
+    @Override
+    public void execute() throws DukeException {
+        System.out.println("    Bye. Hope to see you again soon!");
+    }
+}
