+++ src/main/java/task/Deadline.java
@@ -1,3 +1,4 @@
+package task;
 public class Deadline extends Task {
 
     private String endDate;
