+++ src/main/java/task/Event.java
@@ -1,3 +1,4 @@
+package task;
 public class Event extends Task{
     private String startDate;
     private String endDate;
