+++ src/main/java/task/Task.java
@@ -1,3 +1,4 @@
+package task;
 public class Task {
     private boolean isCompleted = false; //by default the task should not be completed
     private String task;
